refactor(tests): use Swift Testing instead of Quick and Nimble
