+++ Package.swift
@@ -23,8 +23,6 @@ let package = Package(
 		),
 	],
 	dependencies: [
-		.package(url: "https://github.com/Quick/Quick.git", from: "4.0.0"),
-		.package(url: "https://github.com/Quick/Nimble.git", from: "9.0.0"),
 	],
 	targets: [
 		.target(
@@ -53,8 +51,6 @@ let package = Package(
 			name: "FueledUtilsCombineTests",
 			dependencies: [
 				"FueledUtilsCombine",
-				"Quick",
-				"Nimble",
 			],
 			path: "Tests/FueledUtils/CombineTests"
 		),
