+++ Tests/FueledUtils/CombineTests/SinkForLifetimeTests.swift
@@ -1,0 +1,46 @@
+// Copyright © 2024 Fueled Digital Media, LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+@testable import FueledUtilsCombine
+
+import Combine
+import Foundation
+import Testing
+
+@Suite("Sink For Life Time")
+struct SinkForLifetimeTests {
+	@Test("Should cancel itself automatically when the object becomes out of scope")
+	func whenOutOfScope() async throws {
+		var object: NSObject!
+		var valueCount = 0
+		var cancelCount = 0
+			object = NSObject()
+		Timer.publish(every: 0.42, on: .main, in: .common)
+				.autoconnect()
+				.handleEvents(
+					receiveCancel: {
+						cancelCount += 1
+					}
+				)
+				.sinkForLifetimeOf(object) { _ in
+					valueCount += 1
+				}
+
+		try await Task.sleep(for: .seconds(1))
+		object = nil
+
+		#expect(valueCount == 2)
+		#expect(cancelCount == 1)
+	}
+}
